requirements.txt
PyYAML>=6.0
MetaTrader5==5.0.5120
pandas>=2.0
requests>=2.25
websocket-client>=1.5
python-dotenv>=1.0
pytz>=2023.3
pytest>=7.0
